feat: add context API to the app
